+++ src/DriverTracker.h
@@ -8,12 +8,13 @@
 
 class Lap;
 class Stint;
+class Race;
 
 
 class DriverTracker : public F1PacketInterface
 {
   public:
-	DriverTracker(int driverIndex = 0);
+	DriverTracker(int driverIndex = 0, bool raceOnly = false);
 	virtual ~DriverTracker() override {}
 
 	void init(const QDir &directory);
@@ -33,6 +34,7 @@ class DriverTracker : public F1PacketInterface
   protected:
 	bool _extendedPlayerTelemetry = false;
 	int _driverIndex;
+	bool _raceOnly;
 	bool _isPlayer = false;
 	QDir dataDirectory;
 	bool driverDirDefined = false;
@@ -42,13 +44,20 @@ class DriverTracker : public F1PacketInterface
 	CarStatusData _currentStatusData;
 	PacketMotionData _currentMotionData;
 	bool _isLapRecorded = false;
+	bool _sessionFlashbackOccured = false;
 	Lap *_currentLap = nullptr;
 	Stint *_currentStint = nullptr;
+	Race *_currentRace = nullptr;
 	int _currentLapNum = 1;
 	int _currentStintNum = 1;
 	double _startErsBalance = 0;
 	double _timeDiff = 0;
 	double _previousTelemetryDistance = 0;
+	bool _raceFinished = false;
+
+	float startPitTime = 0;
+
+	int lastRaceSessionPassedTime = 0;
 
 	TyresData<DegradationData> degradations;
 
@@ -56,15 +65,29 @@ class DriverTracker : public F1PacketInterface
 	bool flashbackDetected(const LapData &data) const;
 	double averageTyreWear(const CarStatusData &carStatus) const;
 
+	bool isRace() const;
 	bool isLastRaceLap(const LapData &data) const;
 
 	void saveCurrentStint();
 	void addLapToStint(Lap *lap);
 
+	void addLapToRace(Lap *lap, const LapData &lapData);
+	void recordRaceStint(const LapData &lapData);
+	void recordRaceLapEvents(const LapData &lapData);
+	void saveCurrentRace();
+
+	void initLap(Lap *lap, const LapData &lapData);
 	void saveCurrentLap(const LapData &lapData);
 	void startLap(const LapData &lapData);
 
 	void makeDriverDir();
+
+	int sessionTimePassed();
+
+	LapData leaderLapData(const PacketLapData &lapsData) const;
+
+  private slots:
+	void onSessionEnd();
 };
 
 #endif // DRIVERTRACKER_H