+++ src/TelemetryData.cpp
@@ -7,6 +7,8 @@
 
 TelemetryData::TelemetryData(const QVector<TelemetryInfo> &dataInfo) : _telemetryInfo(dataInfo) {}
 
+QVariant TelemetryData::autoSortData() const { return description(); }
+
 void TelemetryData::addData(float x, const QVector<float> &dataValues)
 {
 	_xValues.append(x);
@@ -25,8 +27,10 @@ int TelemetryData::countData() const { return _data.count(); }
 
 void TelemetryData::removeLastData()
 {
-	_xValues.removeLast();
-	_data.removeLast();
+	if(hasData()) {
+		_xValues.removeLast();
+		_data.removeLast();
+	}
 }
 
 QVector<float> TelemetryData::xValues() const { return _xValues; }
@@ -42,6 +46,8 @@ QVector<float> TelemetryData::data(int index) const
 
 	return dataValues;
 }
+
+QVector<float> TelemetryData::lastRecordedData() const { return _data.last(); }
 
 void TelemetryData::setTelemetryInfo(const QVector<TelemetryInfo> &dataNames) { _telemetryInfo = dataNames; }