+++ Changelog.md
@@ -1,5 +1,20 @@
 ## PXG F1 Telemetry Changelog
 
+
+### Version 1.5
+
+- Adds a new "Race Analysis" tab. New files are produced for each driver during a race for each driver. These files can be opened in the race analysis tab to display some statistic on the race
+- Adds a new option to record race data for all drivers, even those which are not tracked
+- Adds fuel mix and ers telemetry data on a lap
+- Adds the possibility to change the theme of the telemetry charts from the menu "View > Theme"
+- Highlights the line of the selected data
+- Adds the possibility to change the color of a single line in the chart
+- LAST VERSION FOR F1 2019
+
+#### Known issues:
+
+- In time trial the telemetry data of ghosts are not properly recorded (this is an issue with the game)
+
 ### Version 1.4
 
 - Allows settings the listened port and IP address.
@@ -11,10 +26,6 @@
 - Improves the readability of the turn names in the graphs
 - Adds a feedback dialog in the help menu
 - Fixes the recording of the player data in time trial
-
-#### Known issues:
-
-- In time trial the telemetry data of ghosts are not properly recorded (this is an issue with the game)
 
 
 ### Version 1.3